+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Updated
   * Updated terra list examples.
+* Changed
+  * Updated documentation for `terra-form-select`.
 
 ## 1.65.0 - (March 8, 2024)